feat(package.json):  make "embrasure" command available in terminal
First, in root directory, run command `npm link`

This makes bin field in package.json  expose module in bin folder to be available globally in the terminal.

Next, we can type "embrasure" in terminal and the module in bin folder will be executed.

https://medium.com/@alexishevia/the-magic-behind-npm-link-d94dcb3a81af

https: //medium.com/netscape/a-guide-to-create-a-nodejs-command-line-package-c2166ad0452e
Co-Authored-By: Taron <88250538+taronavagyan@users.noreply.github.com>